Project27_Plant on C (Bogdan Bakun)
Looks like projects is done. There are a lot of crutches bcs it's too difficult to write on clear C, but it works.
There txt files with tests.
I tried to find out how to write makefile, but what i found was too difficult for me to understand. "Project_27.cbp" opens project in my IDE.